refactor(cli): one XML escaper for the ship renderers
`mime.ts` and `msi.ts` each grew their own — five replacements, the same
entities in the same order, byte-identical bodies — one milestone apart, by
people who never saw the other file. `diff` over the two function bodies is
empty. `utils/ship/xml.ts` is the one home; the second copy is where this
repository's rule says to lift the helper.

`plist.ts` keeps its own three-entity version and that divergence CARRIES
something: a plist value lands only in element CONTENT, so `"` and `'` are
deliberately left as themselves rather than becoming entities a human reading
`Info.plist` has to decode. The new module's header says so, which is what stops
the next reader merging that one too.

NO BEHAVIOUR: the moved body is byte-identical, and both call sites already had
coverage — `msi.spec.ts` asserts the escape directly, `mime.spec.ts` asserts the
rendered document. 3455 unit tests before, 3455 after.
